requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai>=1.61.1
langchain>=0.3.0
langchain-openai>=0.3.3
langchain-community>=0.3.0
langchain-core>=0.3.0
pydantic>=2.7.4,<3.0.0
python-multipart==0.0.9
tiktoken>=0.8.0
unstructured>=0.10.30
markdown>=3.4.3
python-magic>=0.4.27
pyyaml>=6.0.1
firebase-admin>=6.4.0
faiss-cpu>=1.7.4
cohere>=5.0.0
bl